打可部署jar包 框架升级

+++ ev-gb-gateway-server/src/main/java/com/dyy/tsp/evgb/gateway/server/enumtype/GatewayCoreType.java
@@ -68,7 +68,7 @@ public enum GatewayCoreType {
 
     public static GatewayCoreType valuesOf(Short id) {
         for (GatewayCoreType enums : GatewayCoreType.values()) {
-            if (enums.getId()==id) {
+            if (enums.getId()==id.shortValue()) {
                 return enums;
             }
         }
